+++ app/Http/Controllers/Admin/Ticket/TicketPriorityController.php
@@ -17,6 +17,7 @@ class TicketPriorityController extends Controller
      */
     public function index()
     {
+        $this->authorize('index', TicketPriority::class);
         $priorities = TicketPriority::simplePaginate(15);
         return view('admin.ticket.priority.index', compact('priorities'));
     }
@@ -29,6 +30,7 @@ class TicketPriorityController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', TicketPriority::class);
         return view('admin.ticket.priority.create');
     }
 
@@ -40,20 +42,10 @@ class TicketPriorityController extends Controller
      */
     public function store(TicketPriorityRequest $request)
     {
+        $this->authorize('create', TicketPriority::class);
         $inputs = $request->all();
         $result = TicketPriority::create($inputs);
         return redirect()->route('admin.ticket.priority.index')->with('alertify-success', 'اولویت بندی جدید با موفقیت اضافه شد.');
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Ticket $ticket)
-    {
-        return view('admin.ticket.show', compact('ticket'));
     }
 
     /**
@@ -64,6 +56,7 @@ class TicketPriorityController extends Controller
      */
     public function edit(TicketPriority $priority)
     {
+        $this->authorize('update', TicketPriority::class);
         return view('admin.ticket.priority.edit', compact('priority'));
     }
 
@@ -76,6 +69,7 @@ class TicketPriorityController extends Controller
      */
     public function update(TicketPriorityRequest $request, TicketPriority $priority)
     {
+        $this->authorize('update', TicketPriority::class);
         $inputs = $request->all();
         $priority->update($inputs);
         return redirect()->route('admin.ticket.priority.index')->with('alertify-warning', 'اولویت بندی موردنظر با موفقیت ویرایش شد.');
@@ -89,12 +83,14 @@ class TicketPriorityController extends Controller
      */
     public function destroy(TicketPriority $priority)
     {
+        $this->authorize('delete', TicketPriority::class);
         $priority->delete();
         return redirect()->route('admin.ticket.priority.index')->with('alertify-error', 'اولویت بندی موردنظر با موفقیت حذف شد.');
     }
 
     public function status(TicketPriority $priority)
     {
+        $this->authorize('update', TicketPriority::class);
         $priority->status = $priority->status == 0 ? 1 : 0;
         $result = $priority->save();
